+++ scripts/extensions/rollResolver.js
@@ -59,6 +59,16 @@ async function evaluate(wrapped, options) {
     this._total = remoteRoll._total;
     return this;
 }
+async function build(wrapped, config={}, dialog={}, message={}) {
+    let rollType = genericUtils.getProperty(message, 'data.flags.dnd5e.roll.type');
+    if (rollType) {
+        for (let roll of config.rolls ?? []) {
+            roll.options ??= {};
+            roll.options.type ??= rollType;
+        }
+    }
+    return await wrapped(config, dialog, message);
+}
 function patch(enabled) {
     if (enabled) {
         genericUtils.log('dev', 'Evaluate Roll Patched!');
@@ -68,9 +78,20 @@ function patch(enabled) {
         libWrapper.unregister('chris-premades', 'Roll.prototype.evaluate');
     }
 }
+// Update this for 4.2 when change, remove when able
+function patchBuild(enabled) {
+    if (enabled) {
+        genericUtils.log('dev', 'Build D20Roll Patched!');
+        libWrapper.register('chris-premades', 'CONFIG.Dice.D20Roll.build', build, 'WRAPPER');
+    } else {
+        genericUtils.log('dev', 'Build D20Roll Patch Removed!');
+        libWrapper.unregister('chris-premades', 'CONFIG.Dice.D20Roll.build');
+    }
+}
 export let rollResolver = {
     registerFulfillmentMethod,
     unregisterFulfillmentMethod,
     manualRollsUsersDialog,
-    patch
+    patch,
+    patchBuild
 };