+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.1.16 Release:
 ## Bug Fixes:
 - BG3 Weapon Actions
+- Manual Rolling
 ## Update Notes:
 - Registered CPR fields with DAE more cleanly. Requires DAE 12.0.10+